requirements.txt
asgiref==3.7.2
Django==5.0.2
django-froala-editor==4.1.4
django-js-asset==2.2.0
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
